calculations/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
heartpy==1.2.6
idna==2.9
influxdb==5.3.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.0.0
msgpack==0.6.1
numpy==1.18.4
pandas==0.25.3
patsy==0.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
statsmodels==0.11.1
urllib3==1.25.9
